prepopulate database

+++ app/src/main/java/com/dicoding/habitapp/data/Habit.kt
@@ -18,7 +18,7 @@ data class Habit(
     val title: String,
 
     @ColumnInfo(name="minutesFocus")
-    val minutesFocus: Int,
+    val minutesFocus: Long,
 
     @ColumnInfo(name="startTime")
     val startTime: String,
